+++ src/main/java/com/pocalink/ddingcon/domain/auth/controller/AuthController.java
@@ -1,0 +1,23 @@
+package com.pocalink.ddingcon.domain.auth.controller;
+
+import com.pocalink.ddingcon.global.security.AuthTokens;
+import com.pocalink.ddingcon.domain.auth.service.OAuthLoginService;
+import com.pocalink.ddingcon.global.security.KakaoLoginParams;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/auth")
+public class AuthController {
+    private final OAuthLoginService oAuthLoginService;
+
+    @PostMapping("/kakao")
+    public ResponseEntity<AuthTokens> loginKakao(@RequestBody KakaoLoginParams params) {
+        return ResponseEntity.ok(oAuthLoginService.login(params));
+    }
+}
