+++ src/main/java/com/pocalink/ddingcon/domain/auth/service/KakaoApiClient.java
@@ -3,8 +3,8 @@ package com.pocalink.ddingcon.domain.auth.service;
 import com.pocalink.ddingcon.domain.auth.dto.response.KakaoInfoResponse;
 import com.pocalink.ddingcon.domain.auth.dto.response.KakaoTokens;
 import com.pocalink.ddingcon.domain.auth.dto.response.OAuthInfoResponse;
-import com.pocalink.ddingcon.domain.member.domain.OAuthLoginParams;
-import com.pocalink.ddingcon.domain.member.domain.OAuthProvider;
+import com.pocalink.ddingcon.global.security.OAuthLoginParams;
+import com.pocalink.ddingcon.global.security.OAuthProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
