+++ src/main/java/com/pocalink/ddingcon/domain/auth/service/OAuthApiClient.java
@@ -1,8 +1,8 @@
 package com.pocalink.ddingcon.domain.auth.service;
 
 import com.pocalink.ddingcon.domain.auth.dto.response.OAuthInfoResponse;
-import com.pocalink.ddingcon.domain.member.domain.OAuthLoginParams;
-import com.pocalink.ddingcon.domain.member.domain.OAuthProvider;
+import com.pocalink.ddingcon.global.security.OAuthLoginParams;
+import com.pocalink.ddingcon.global.security.OAuthProvider;
 
 public interface OAuthApiClient {
     OAuthProvider oAuthProvider();
