+++ src/main/java/com/pocalink/ddingcon/domain/member/controller/MemberController.java
@@ -1,0 +1,32 @@
+package com.pocalink.ddingcon.domain.member.controller;
+
+import com.pocalink.ddingcon.domain.member.domain.Member;
+import com.pocalink.ddingcon.domain.member.repository.MemberRepository;
+import com.pocalink.ddingcon.global.security.AuthTokensGenerator;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/members")
+public class MemberController {
+    private final MemberRepository memberRepository;
+    private final AuthTokensGenerator authTokensGenerator;
+
+    @GetMapping
+    public ResponseEntity<List<Member>> findAll() {
+        return ResponseEntity.ok(memberRepository.findAll());
+    }
+
+    @GetMapping("/{accessToken}")
+    public ResponseEntity<Member> findByAccessToken(@PathVariable String accessToken) {
+        Long memberId = authTokensGenerator.extractMemberId(accessToken);
+        return ResponseEntity.ok(memberRepository.findById(memberId).get());
+    }
+}
