+++ src/main/java/com/pocalink/ddingcon/domain/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.pocalink.ddingcon.domain.member.repository;
+
+import com.pocalink.ddingcon.domain.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByNickname(String nickname);
+}
