--- src/main/java/com/pocalink/ddingcon/global/config/ClientConfig.java
@@ -1,13 +1,0 @@
-package com.pocalink.ddingcon.global.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class ClientConfig {
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
