+++ src/main/java/com/pocalink/ddingcon/global/security/JwtTokenProvider.java
@@ -1,64 +1,49 @@
 package com.pocalink.ddingcon.global.security;
 
-import io.jsonwebtoken.*;
-import jakarta.servlet.http.HttpServletRequest;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.security.Key;
 import java.util.Date;
-
-import static io.jsonwebtoken.security.Keys.hmacShaKeyFor;
-import static java.nio.charset.StandardCharsets.UTF_8;
 
 @Component
 public class JwtTokenProvider {
 
-    @Value("${jwt.secret}")
-    private String secretKey;
-    @Value("${jwt.accesstoken-validity-in-seconds}")
-    private long accessTokenExpirationTime;
-    @Value("${jwt.refreshtoken-validity-in-seconds}")
-    private long refreshTokenExpirationTime;
+    private final Key key;
 
-    public String createAccessToken(Long memberId) {
-        return createToken(memberId.toString(), accessTokenExpirationTime);
+    public JwtTokenProvider(@Value("${jwt.secret-key}") String secretKey) {
+        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
+        this.key = Keys.hmacShaKeyFor(keyBytes);
     }
 
-    public String createRefreshToken(Long memberId) {
-        return createToken(memberId.toString(), refreshTokenExpirationTime);
-    }
-
-    public String createToken(String payload, long expirationTime) {
-        Date now = new Date();
+    public String generate(String subject, Date expiredAt) {
         return Jwts.builder()
-                .setSubject(payload)
-                .setIssuedAt(now)
-                .setExpiration(new Date(now.getTime() + expirationTime))
-                .signWith(hmacShaKeyFor(secretKey.getBytes(UTF_8)), SignatureAlgorithm.HS256)
+                .setSubject(subject)
+                .setExpiration(expiredAt)
+                .signWith(key, SignatureAlgorithm.HS512)
                 .compact();
     }
 
-    public String extractToken(HttpServletRequest request) {
-        String token = request.getHeader("Authorization");
-        if (token != null && token.startsWith("Bearer ")) {
-            return token.substring("Bearer ".length());
-        }
-        return null;
+    public String extractSubject(String accessToken) {
+        Claims claims = parseClaims(accessToken);
+        return claims.getSubject();
     }
 
-    public boolean validateToken(String token) {
+    private Claims parseClaims(String accessToken) {
         try {
-            Jws<Claims> claims = Jwts.parserBuilder()
-                    .setSigningKey(secretKey.getBytes(UTF_8))
+            return Jwts.parserBuilder()
+                    .setSigningKey(key)
                     .build()
-                    .parseClaimsJws(token);
-            return claims.getBody()
-                    .getExpiration()
-                    .after(new Date());
+                    .parseClaimsJws(accessToken)
+                    .getBody();
         } catch (ExpiredJwtException e) {
-            return false;
-        } catch (JwtException e) {
-            return false;
+            return e.getClaims();
         }
     }
 }
