+++ src/main/java/com/pocalink/ddingcon/global/security/OAuthProvider.java
@@ -1,4 +1,4 @@
-package com.pocalink.ddingcon.domain.member.domain;
+package com.pocalink.ddingcon.global.security;
 
 
 // 로그인 인증 플랫폼 추가를 고려해 ENUM 생성
